requirements.txt
cffi==1.9.1
click==6.6
Flask==0.11.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pycparser==2.17
PyNaCl==1.0.1
six==1.10.0
Werkzeug==0.11.11
